refactor(test) Rename systemmodel to composermodel in tests
