+++ packages/concerto-core/test/composer/composermodel.js
@@ -14,8 +14,7 @@
 
 'use strict';
 
-const SYSTEM_MODELS = [
-
+const COMPOSER_MODEL =
     {
         namespace: 'org.hyperledger.composer.system',
         fileName: 'org.hyperledger.composer.system.cto',
@@ -39,7 +38,7 @@ namespace org.hyperledger.composer.system
  * Has no properties, and is soley used as a basis to model other assets.
  */
 @docs('participant.md')
-abstract participant HLParticipant identified by participantId {
+abstract participant ComposerParticipant identified by participantId {
   o String participantId
 }
 
@@ -123,13 +122,13 @@ participant NetworkAdmin identified by participantId {
 @docs('historian.md')
 @docsuri('Composer Documentation','../business-network/historian')
 asset HistorianRecord identified by transactionId {
-  o String          transactionId
-  o String          transactionType
-  --> Transaction   transactionInvoked
-  --> HLParticipant participantInvoking  optional
-  --> Identity      identityUsed         optional
-  o Event[]         eventsEmitted        optional
-  o DateTime        transactionTimestamp
+  o String                transactionId
+  o String                transactionType
+  --> Transaction         transactionInvoked
+  --> ComposerParticipant participantInvoking  optional
+  --> Identity            identityUsed         optional
+  o Event[]               eventsEmitted        optional
+  o DateTime              transactionTimestamp
 }
 
 // -----------------------------------------------------------------------------
@@ -232,7 +231,7 @@ asset Identity identified by identityId {
     o String issuer
     o String certificate
     o IdentityState state
-    --> HLParticipant participant
+    --> ComposerParticipant participant
 }
 
 /**
@@ -242,7 +241,7 @@ asset Identity identified by identityId {
  */
 @docs('issueIdentity.md')
 transaction IssueIdentity {
-    --> HLParticipant participant
+    --> ComposerParticipant participant
     o String identityName
 }
 
@@ -253,7 +252,7 @@ transaction IssueIdentity {
  */
 @docs('bindIdentity.md')
 transaction BindIdentity {
-    --> HLParticipant participant
+    --> ComposerParticipant participant
     o String certificate
 }
 
@@ -302,7 +301,6 @@ transaction SetLogLevel {
 }
 `
     }
+;
 
-];
-
-module.exports = SYSTEM_MODELS;
+module.exports = COMPOSER_MODEL;
