+++ packages/concerto-core/test/composer/composermodelutility.js
@@ -14,18 +14,16 @@
 
 'use strict';
 
-const debug = require('debug')('concerto-core:systemmodelutility');
+const debug = require('debug')('concerto-core:composermodelutility');
 
-const SYSTEM_MODELS = require('./systemmodel');
+const COMPOSER_MODEL = require('./composermodel');
 
 module.exports = {
-    addComposerSystemModels: (modelManager) => {
-        SYSTEM_MODELS.forEach((SYSTEM_MODEL) => {
-            debug('addComposerSystemModels', SYSTEM_MODEL);
-            modelManager.addModelFile(SYSTEM_MODEL.contents, SYSTEM_MODEL.fileName, true);
-        });
+    addComposerModel: (modelManager) => {
+        debug('addComposerModel', COMPOSER_MODEL);
+        modelManager.addModelFile(COMPOSER_MODEL.contents, COMPOSER_MODEL.fileName, true);
 
-        // now validate all the models
+        // now validate the model
         modelManager.validateModelFiles();
     },
 };
