+++ packages/concerto-core/test/composer/i18n.js
@@ -23,7 +23,7 @@ const Factory = require('./../../lib/factory');
 const Serializer = require('./../../lib/serializer');
 const ModelManager = require('./../../lib/modelmanager');
 const ModelUtil = require('./../../lib/modelutil');
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 const Moment = require('moment-mini');
 
 describe('Globalization', function() {
@@ -113,7 +113,7 @@ describe('Globalization', function() {
 
                 // create and polulate the modelManager with a model file
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 let fileName = './test/composer/models/classdeclaration/validate/foo-identifiernotproperty.cto';
                 let invalidFile = fs.readFileSync(fileName, 'utf8');
@@ -133,7 +133,7 @@ describe('Globalization', function() {
 
                 // create and polulate the modelManager with a model file
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 let fileName = './test/composer/models/classdeclaration/validate/foo-identifiernotstring.cto';
                 let invalidFile = fs.readFileSync(fileName, 'utf8');
@@ -164,7 +164,7 @@ describe('Globalization', function() {
 
             // create and polulate the modelManager with a model file
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
 
             let fileName = './test/composer/models/modelfile/resolvetype/foo-undecltype.cto';
             let invalidFile = fs.readFileSync(fileName, 'utf8');
@@ -203,7 +203,7 @@ describe('Globalization', function() {
                 }).should.equal('ModelFile for namespace foo has not been registered with the ModelManager');
 
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 expect(function() {
                     let factory = new Factory(modelManager);
@@ -219,7 +219,7 @@ describe('Globalization', function() {
                 }).should.equal('Type bar is not declared in namespace foo');
 
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 let fileName = './test/composer/models/factory/newinstance/foo-typenotdeclaredinns.cto';
                 let file = fs.readFileSync(fileName, 'utf8');
@@ -256,7 +256,7 @@ describe('Globalization', function() {
                 }).should.equal('Type bar is not declared in namespace foo');
 
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 let fileName = './test/composer/models/factory/newinstance/foo-typenotdeclaredinns.cto';
                 let file = fs.readFileSync(fileName, 'utf8');
@@ -278,7 +278,7 @@ describe('Globalization', function() {
                 }).should.equal('ModelFile for namespace foo has not been registered with the ModelManager');
 
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 expect(function() {
                     let factory = new Factory(modelManager);
@@ -294,7 +294,7 @@ describe('Globalization', function() {
                 }).should.equal('Type bar is not declared in namespace foo');
 
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 let fileName = './test/composer/models/factory/newrelationship/foo-typenotdeclaredinns.cto';
                 let file = fs.readFileSync(fileName, 'utf8');
@@ -319,7 +319,7 @@ describe('Globalization', function() {
                 }).should.equal('No registered namespace for type bar in foo');
 
                 const modelManager = new ModelManager();
-                Util.addComposerSystemModels(modelManager);
+                Util.addComposerModel(modelManager);
 
                 expect(function() {
                     modelManager.resolveType('foo', 'bar');
@@ -370,7 +370,7 @@ describe('Globalization', function() {
             formatter().should.equal('Factory cannot be null');
 
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             expect(function() {
                 new Serializer(null, modelManager);
             }).to.throw(Error, 'Factory cannot be null');
@@ -381,7 +381,7 @@ describe('Globalization', function() {
             formatter().should.equal('Serializer.toJSON only accepts Concept, Event, Asset, Participant or Transaction.');
 
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             expect(function() {
                 let serializer = new Serializer(true, modelManager);
                 serializer.toJSON({});
