+++ packages/concerto-core/test/introspect/eventdeclaration.js
@@ -17,7 +17,7 @@
 const EventDeclaration = require('../../lib/introspect/eventdeclaration');
 const ModelManager = require('../../lib/modelmanager');
 const fs = require('fs');
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 
 require('chai').should();
 const sinon = require('sinon');
@@ -51,7 +51,7 @@ describe('EventDeclaration', () => {
                 o String euid
             }`;
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             const modelFile = modelManager.addModelFile(model, 'awesome.cto' );
 
             let ed = modelFile.getEventDeclaration('E');
@@ -70,7 +70,7 @@ describe('EventDeclaration', () => {
             const fileName = 'test/data/model/events.cto';
             let modelDefinitions = fs.readFileSync(fileName, 'utf8');
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             const modelFile = modelManager.addModelFile(modelDefinitions, fileName );
 
             const abstractEvent = modelFile.getEventDeclaration('AbstractEvent');
