+++ packages/concerto-core/test/introspect/introspector.js
@@ -16,7 +16,7 @@
 
 const ModelManager = require('../../lib/modelmanager');
 const Introspector = require('../../lib/introspect/introspector');
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 
 const fs = require('fs');
 
@@ -46,7 +46,7 @@ describe('Introspector', () => {
         it('should return all class declarations', () => {
             // create and populate the ModelManager with a model file
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             modelManager.should.not.be.null;
 
             let modelBase = fs.readFileSync('./test/data/model/model-base.cto', 'utf8');
@@ -66,7 +66,7 @@ describe('Introspector', () => {
         it('should be able to get a single class declaration', () => {
             // create and populate the ModelManager with a model file
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             modelManager.should.not.be.null;
 
             let modelBase = fs.readFileSync('./test/data/model/model-base.cto', 'utf8');
@@ -82,7 +82,7 @@ describe('Introspector', () => {
 
         it('should return the model manager', () => {
             const modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             const introspector = new Introspector(modelManager);
             introspector.getModelManager().should.equal(modelManager);
         });
