+++ packages/concerto-core/test/introspect/participantdeclaration.js
@@ -17,7 +17,7 @@
 const ModelFile = require('../../lib/introspect/modelfile');
 const ModelManager = require('../../lib/modelmanager');
 const fs = require('fs');
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 
 require('chai').should();
 
@@ -27,7 +27,7 @@ describe('ParticipantDeclaration', () => {
 
     beforeEach(() => {
         modelManager = new ModelManager();
-        Util.addComposerSystemModels(modelManager);
+        Util.addComposerModel(modelManager);
     });
 
     let loadParticipantDeclaration = (modelFileName) => {
