+++ packages/concerto-core/test/model/identifiable.js
@@ -20,7 +20,7 @@ const sinon = require('sinon');
 const chai = require('chai');
 chai.should();
 chai.use(require('chai-things'));
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 
 describe('Identifiable', function () {
 
@@ -28,7 +28,7 @@ describe('Identifiable', function () {
     let classDecl;
     before(function () {
         modelManager = new ModelManager();
-        Util.addComposerSystemModels(modelManager);
+        Util.addComposerModel(modelManager);
         modelManager.addModelFile(`namespace com.composer
         participant Farmer identified by farmerId {
             o String farmerId
