+++ packages/concerto-core/test/model/relationship.js
@@ -16,7 +16,7 @@
 
 const ModelManager = require('../../lib/modelmanager');
 const Relationship = require('../../lib/model/relationship');
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 
 const chai = require('chai');
 chai.should();
@@ -39,7 +39,7 @@ describe('Relationship', function () {
 
     before(function () {
         modelManager = new ModelManager();
-        Util.addComposerSystemModels(modelManager);
+        Util.addComposerModel(modelManager);
     });
 
     beforeEach(function () {
