+++ packages/concerto-core/test/models/farm2fork.js
@@ -17,7 +17,7 @@
 require('chai').should();
 const ModelManager = require('../../lib/modelmanager');
 const RelationshipDeclaration = require('../../lib/introspect/relationshipdeclaration');
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 
 const fs = require('fs');
 
@@ -25,7 +25,7 @@ describe('Farm2Fork Model', function() {
     describe('#getModelManagerFarm2Fork', function() {
         it('check parsing and model manager', function() {
             let modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             modelManager.should.not.be.null;
 
             // parse a model file from disk and add to the ModelManager
