+++ packages/concerto-core/test/models/wildcards.js
@@ -18,7 +18,7 @@ const Factory = require('../../lib/factory');
 const fs = require('fs');
 const ModelManager = require('../../lib/modelmanager');
 const Serializer = require('../../lib/serializer');
-const Util = require('../composer/systemmodelutility');
+const Util = require('../composer/composermodelutility');
 
 require('chai').should();
 
@@ -31,7 +31,7 @@ describe('Wildcards Model', function () {
 
     beforeEach(() => {
         modelManager = new ModelManager();
-        Util.addComposerSystemModels(modelManager);
+        Util.addComposerModel(modelManager);
         factory = new Factory(modelManager);
         serializer = new Serializer(factory, modelManager);
         ergoSerializer = new Serializer(factory, modelManager);
