+++ packages/concerto-core/test/modelutil.js
@@ -19,7 +19,7 @@ const Property = require('../lib/introspect/property');
 const Typed = require('../lib/model/identifiable');
 const ModelManager = require('../lib/modelmanager');
 const ModelUtil = require('../lib/modelutil');
-const Util = require('./composer/systemmodelutility');
+const Util = require('./composer/composermodelutility');
 
 require('chai').should();
 const sinon = require('sinon');
@@ -115,7 +115,7 @@ describe('ModelUtil', function () {
 
         beforeEach(() => {
             modelManager = new ModelManager();
-            Util.addComposerSystemModels(modelManager);
+            Util.addComposerModel(modelManager);
             modelManager.addModelFile(`namespace org.acme.baz
             asset Foo identified by fooId {
                 o String fooId
